refactor slot hash
